feat: chat profiles and knowledge bases
Knowledge bases:
Knowledge bases are basically a place for grouping files that are related and thereafter be attached to the chat profiles.
Files can be added to the knowledge bases. When a chat is created from a chat profile with knowledge bases. The files in
the knowledge bases will be included in the new chat and become the context.

Chat profiles:
A chat profile is essentially a pre-configuration contains the files, advanced params, system prompt for a conversation.
Chats created from a chat profile will have everything defined in the profile so users don't need to configure the params
and attach the files manually. If needed, we can disable the creation of the normal new chats and only expose the chat profiles
so all new chats can only be created from designated profiles.